+++ src/ecanvasser/services/survey.service.ts
@@ -1,28 +1,22 @@
-import { Injectable, Logger, NotFoundException } from '@nestjs/common'
-import { HttpService } from '@nestjs/axios'
+import { Injectable, NotFoundException } from '@nestjs/common'
 import { EcanvasserService } from './ecanvasser.service'
-import { EcanvasserApiUtil } from '../util/ecanvasser.util'
+import { CreateSurveySchema } from '../dto/createSurvey.schema'
+import { CreateSurveyQuestionSchema } from '../dto/createSurveyQuestion.schema'
+import { UpdateSurveyQuestionSchema } from '../dto/updateSurveyQuestion.schema'
+import { UpdateSurveySchema } from '../dto/updateSurvey.schema'
 import {
   ApiEcanvasserSurvey,
   ApiEcanvasserSurveyQuestion,
   ApiEcanvasserTeam,
 } from '../ecanvasser.types'
-import { CreateSurveySchema } from '../dto/createSurvey.schema'
-import { CreateSurveyQuestionSchema } from '../dto/createSurveyQuestion.schema'
-import { UpdateSurveyQuestionSchema } from '../dto/updateSurveyQuestion.schema'
-import { UpdateSurveySchema } from '../dto/updateSurvey.schema'
+import { EcanvasserApiService } from './ecanvasserAPI.service'
 
 @Injectable()
 export class SurveyService {
-  private readonly logger = new Logger(SurveyService.name)
-  private readonly apiUtil: EcanvasserApiUtil
-
   constructor(
-    private readonly httpService: HttpService,
     private readonly ecanvasserService: EcanvasserService,
-  ) {
-    this.apiUtil = new EcanvasserApiUtil(httpService)
-  }
+    private readonly ecanvasserApi: EcanvasserApiService,
+  ) {}
 
   async createSurvey(campaignId: number, createSurveyDto: CreateSurveySchema) {
     const ecanvasser = await this.ecanvasserService.findByCampaignId(campaignId)
@@ -36,23 +30,18 @@ export class SurveyService {
       requires_signature: createSurveyDto.requiresSignature,
       status: createSurveyDto.status,
       team_id: createSurveyDto.teamId,
-    } as ApiEcanvasserSurvey
-
-    try {
-      const response = await this.apiUtil.fetchFromApi<ApiEcanvasserSurvey>(
-        '/survey',
-        ecanvasser.apiKey,
-        {
-          method: 'POST',
-          data: payload,
-        },
-      )
-
-      return response.data
-    } catch (error) {
-      this.logger.error('Failed to create survey', error)
-      throw error
     }
+
+    const response = await this.ecanvasserApi.fetchFromApi<ApiEcanvasserSurvey>(
+      '/survey',
+      ecanvasser.apiKey,
+      {
+        method: 'POST',
+        data: payload,
+      },
+    )
+
+    return response.data
   }
 
   async findSurveys(campaignId: number) {
@@ -61,17 +50,65 @@ export class SurveyService {
       throw new NotFoundException('Ecanvasser integration not found')
     }
 
-    try {
-      const response = await this.apiUtil.fetchFromApi<ApiEcanvasserSurvey>(
-        '/survey',
-        ecanvasser.apiKey,
-      )
+    const response = await this.ecanvasserApi.fetchFromApi<ApiEcanvasserSurvey>(
+      '/survey',
+      ecanvasser.apiKey,
+    )
 
-      return response.data
-    } catch (error) {
-      this.logger.error('Failed to fetch surveys', error)
-      throw error
+    return response.data
+  }
+
+  async findSurvey(campaignId: number, surveyId: number) {
+    const ecanvasser = await this.ecanvasserService.findByCampaignId(campaignId)
+    if (!ecanvasser) {
+      throw new NotFoundException('Ecanvasser integration not found')
     }
+
+    const response = await this.ecanvasserApi.fetchFromApi<ApiEcanvasserSurvey>(
+      `/survey/${surveyId}`,
+      ecanvasser.apiKey,
+    )
+
+    return response.data
+  }
+
+  async updateSurvey(
+    campaignId: number,
+    surveyId: number,
+    updateSurveyDto: UpdateSurveySchema,
+  ) {
+    const ecanvasser = await this.ecanvasserService.findByCampaignId(campaignId)
+    if (!ecanvasser) {
+      throw new NotFoundException('Ecanvasser integration not found')
+    }
+
+    const response = await this.ecanvasserApi.fetchFromApi<ApiEcanvasserSurvey>(
+      `/survey/${surveyId}`,
+      ecanvasser.apiKey,
+      {
+        method: 'PUT',
+        data: updateSurveyDto,
+      },
+    )
+
+    return response.data
+  }
+
+  async deleteSurvey(campaignId: number, surveyId: number) {
+    const ecanvasser = await this.ecanvasserService.findByCampaignId(campaignId)
+    if (!ecanvasser) {
+      throw new NotFoundException('Ecanvasser integration not found')
+    }
+
+    const response = await this.ecanvasserApi.fetchFromApi(
+      `/survey/${surveyId}`,
+      ecanvasser.apiKey,
+      {
+        method: 'DELETE',
+      },
+    )
+
+    return response.data
   }
 
   async createSurveyQuestion(
@@ -96,82 +133,17 @@ export class SurveyService {
       answers: createQuestionDto.answers || undefined,
     }
 
-    try {
-      const response =
-        await this.apiUtil.fetchFromApi<ApiEcanvasserSurveyQuestion>(
-          `/survey/question`,
-          ecanvasser.apiKey,
-          {
-            method: 'POST',
-            data: payload,
-          },
-        )
-
-      return response.data
-    } catch (error) {
-      this.logger.error('Failed to create survey question', error)
-      throw error
-    }
-  }
-
-  async findSurvey(campaignId: number, surveyId: number) {
-    const ecanvasser = await this.ecanvasserService.findByCampaignId(campaignId)
-    if (!ecanvasser) {
-      throw new NotFoundException('Ecanvasser integration not found')
-    }
-
-    try {
-      const response = await this.apiUtil.fetchFromApi<ApiEcanvasserSurvey>(
-        `/survey/${surveyId}`,
-        ecanvasser.apiKey,
-      )
-
-      return response.data
-    } catch (error) {
-      this.logger.error('Failed to fetch survey', error)
-      throw error
-    }
-  }
-
-  async findTeams(campaignId: number) {
-    const ecanvasser = await this.ecanvasserService.findByCampaignId(campaignId)
-    if (!ecanvasser) {
-      throw new NotFoundException('Ecanvasser integration not found')
-    }
-
-    try {
-      const response = await this.apiUtil.fetchFromApi<ApiEcanvasserTeam>(
-        '/team',
-        ecanvasser.apiKey,
-      )
-
-      return response.data
-    } catch (error) {
-      this.logger.error('Failed to fetch teams', error)
-      throw error
-    }
-  }
-
-  async deleteSurveyQuestion(campaignId: number, questionId: number) {
-    const ecanvasser = await this.ecanvasserService.findByCampaignId(campaignId)
-    if (!ecanvasser) {
-      throw new NotFoundException('Ecanvasser integration not found')
-    }
-
-    try {
-      const response = await this.apiUtil.fetchFromApi(
-        `/survey/question/${questionId}`,
+    const response =
+      await this.ecanvasserApi.fetchFromApi<ApiEcanvasserSurveyQuestion>(
+        '/survey/question',
         ecanvasser.apiKey,
         {
-          method: 'DELETE',
+          method: 'POST',
+          data: payload,
         },
       )
 
-      return response.data
-    } catch (error) {
-      this.logger.error('Failed to delete survey question', error)
-      throw error
-    }
+    return response.data
   }
 
   async findSurveyQuestion(campaignId: number, questionId: number) {
@@ -180,18 +152,13 @@ export class SurveyService {
       throw new NotFoundException('Ecanvasser integration not found')
     }
 
-    try {
-      const response =
-        await this.apiUtil.fetchFromApi<ApiEcanvasserSurveyQuestion>(
-          `/survey/question/${questionId}`,
-          ecanvasser.apiKey,
-        )
+    const response =
+      await this.ecanvasserApi.fetchFromApi<ApiEcanvasserSurveyQuestion>(
+        `/survey/question/${questionId}`,
+        ecanvasser.apiKey,
+      )
 
-      return response.data
-    } catch (error) {
-      this.logger.error('Failed to fetch survey question', error)
-      throw error
-    }
+    return response.data
   }
 
   async updateSurveyQuestion(
@@ -210,70 +177,47 @@ export class SurveyService {
       answers: updateQuestionDto.answers,
     }
 
-    try {
-      const response =
-        await this.apiUtil.fetchFromApi<ApiEcanvasserSurveyQuestion>(
-          `/survey/question/${questionId}`,
-          ecanvasser.apiKey,
-          {
-            method: 'PUT',
-            data: payload,
-          },
-        )
-
-      return response.data
-    } catch (error) {
-      this.logger.error('Failed to update survey question', error)
-      throw error
-    }
-  }
-
-  async updateSurvey(
-    campaignId: number,
-    surveyId: number,
-    updateSurveyDto: UpdateSurveySchema,
-  ) {
-    const ecanvasser = await this.ecanvasserService.findByCampaignId(campaignId)
-    if (!ecanvasser) {
-      throw new NotFoundException('Ecanvasser integration not found')
-    }
-
-    try {
-      const response = await this.apiUtil.fetchFromApi<ApiEcanvasserSurvey>(
-        `/survey/${surveyId}`,
+    const response =
+      await this.ecanvasserApi.fetchFromApi<ApiEcanvasserSurveyQuestion>(
+        `/survey/question/${questionId}`,
         ecanvasser.apiKey,
         {
           method: 'PUT',
-          data: updateSurveyDto,
+          data: payload,
         },
       )
 
-      return response.data
-    } catch (error) {
-      this.logger.error('Failed to update survey', error)
-      throw error
-    }
+    return response.data
   }
 
-  async deleteSurvey(campaignId: number, surveyId: number) {
+  async deleteSurveyQuestion(campaignId: number, questionId: number) {
     const ecanvasser = await this.ecanvasserService.findByCampaignId(campaignId)
     if (!ecanvasser) {
       throw new NotFoundException('Ecanvasser integration not found')
     }
 
-    try {
-      const response = await this.apiUtil.fetchFromApi(
-        `/survey/${surveyId}`,
-        ecanvasser.apiKey,
-        {
-          method: 'DELETE',
-        },
-      )
+    const response = await this.ecanvasserApi.fetchFromApi(
+      `/survey/question/${questionId}`,
+      ecanvasser.apiKey,
+      {
+        method: 'DELETE',
+      },
+    )
 
-      return response.data
-    } catch (error) {
-      this.logger.error('Failed to delete survey', error)
-      throw error
+    return response.data
+  }
+
+  async findTeams(campaignId: number) {
+    const ecanvasser = await this.ecanvasserService.findByCampaignId(campaignId)
+    if (!ecanvasser) {
+      throw new NotFoundException('Ecanvasser integration not found')
     }
+
+    const response = await this.ecanvasserApi.fetchFromApi<ApiEcanvasserTeam>(
+      '/team',
+      ecanvasser.apiKey,
+    )
+
+    return response.data
   }
 }
